# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/health/live": {
      "get": {
        "operationId": "HealthController_liveness",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Liveness probe. Does not touch dependencies.",
        "tags": [
          "health"
        ]
      }
    },
    "/health/ready": {
      "get": {
        "operationId": "HealthController_readiness",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Readiness probe. Checks Postgres and Redis.",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "PHONE_ALREADY_EXISTS or EMAIL_ALREADY_EXISTS"
          }
        },
        "summary": "Register a new student account",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS"
          }
        },
        "summary": "Log in with phone or email",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "Rotating: the presented token is consumed. Presenting a consumed token revokes every token in that session chain and returns TOKEN_REUSED.",
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "TOKEN_EXPIRED, TOKEN_INVALID or TOKEN_REUSED"
          }
        },
        "summary": "Exchange a refresh token for a new pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "Idempotent.",
        "operationId": "AuthController_logout_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Session ended"
          }
        },
        "summary": "End one session",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/me": {
      "get": {
        "description": "A null `profile` means academic setup has not been completed; the app routes to onboarding on that.",
        "operationId": "StudentsController_getMe_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "The signed-in student",
        "tags": [
          "student"
        ]
      },
      "patch": {
        "description": "Serves both first-time academic setup and later edits. `subject_ids` replaces the whole selection rather than merging.",
        "operationId": "StudentsController_updateMe_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Create or update the profile",
        "tags": [
          "student"
        ]
      }
    },
    "/api/v1/curriculum": {
      "get": {
        "operationId": "CurriculumController_listCurricula_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List active curricula and the classes they cover",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/subjects": {
      "get": {
        "operationId": "CurriculumController_listSubjects_v1",
        "parameters": [
          {
            "name": "curriculum",
            "required": false,
            "in": "query",
            "schema": {
              "example": "nctb"
            }
          },
          {
            "name": "class_level",
            "required": false,
            "in": "query",
            "schema": {
              "example": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List subjects, optionally filtered by class and curriculum",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/subjects/{id}": {
      "get": {
        "operationId": "CurriculumController_getSubject_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "One subject with its chapters",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/subjects/{id}/chapters": {
      "get": {
        "operationId": "CurriculumController_listChapters_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Chapters of a subject",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/chapters/{id}": {
      "get": {
        "operationId": "CurriculumController_getChapter_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "One chapter with its topics",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/chapters/{id}/topics": {
      "get": {
        "operationId": "CurriculumController_listTopics_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Topics of a chapter",
        "tags": [
          "curriculum"
        ]
      }
    },
    "/api/v1/admin/documents": {
      "post": {
        "description": "Deduplicates on content hash: uploading identical bytes returns the existing document with `deduplicated: true`.",
        "operationId": "AdminController_uploadDocument_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "title",
                  "document_type"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "title": {
                    "type": "string"
                  },
                  "document_type": {
                    "type": "string",
                    "enum": [
                      "UPLOADED",
                      "QUEUED",
                      "EXTRACTING",
                      "CHUNKING",
                      "EMBEDDING",
                      "READY",
                      "FAILED"
                    ]
                  },
                  "subject_id": {
                    "type": "string"
                  },
                  "chapter_id": {
                    "type": "string"
                  },
                  "source": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Upload a document and queue it for ingestion",
        "tags": [
          "admin"
        ]
      },
      "get": {
        "operationId": "AdminController_listDocuments_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List uploaded documents and their processing state",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/documents/{id}/process": {
      "post": {
        "description": "Existing chunks are replaced, not appended to.",
        "operationId": "AdminController_reprocess_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Re-run ingestion for a document",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/knowledge-base/status": {
      "get": {
        "operationId": "AdminController_status_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Corpus and queue health",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/knowledge-base/chunks": {
      "get": {
        "description": "Search chunk text to find which passage produced a given answer — or that none did.",
        "operationId": "AdminController_chunks_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Chunk inspector",
        "tags": [
          "admin"
        ]
      }
    }
  },
  "info": {
    "title": "AI Tutor Bangladesh API",
    "description": "Curriculum-grounded AI tutor for Class 9–10 Science (NCTB).\n\nAll errors share one envelope: `{ error: { code, message, message_bn, request_id } }`. Error codes are enumerated in `@ai-tutor/shared-types`.",
    "version": "0.1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "auth",
      "description": "Registration, login, token rotation"
    },
    {
      "name": "health",
      "description": "Liveness and readiness probes"
    }
  ],
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Tanim Ahmed",
            "minLength": 2,
            "maxLength": 120
          },
          "phone": {
            "type": "string",
            "example": "+8801712345678",
            "description": "Primary identifier. Bangladeshi mobile number in E.164 format."
          },
          "email": {
            "type": "string",
            "example": "student@example.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 128,
            "example": "correct-horse-battery"
          }
        },
        "required": [
          "name",
          "phone",
          "password"
        ]
      },
      "PublicUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "ADMIN"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "SUSPENDED",
              "DELETED"
            ]
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "email",
          "role",
          "status",
          "created_at"
        ]
      },
      "AuthTokensDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "expires_in": {
            "type": "number",
            "description": "Access token lifetime in seconds."
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          }
        },
        "required": [
          "access_token",
          "refresh_token",
          "expires_in",
          "token_type"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/PublicUserDto"
          },
          "profile": {
            "type": "object",
            "nullable": true,
            "description": "Null until the student completes academic setup."
          },
          "tokens": {
            "$ref": "#/components/schemas/AuthTokensDto"
          }
        },
        "required": [
          "user",
          "tokens"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "+8801712345678",
            "description": "Phone number or email address."
          },
          "password": {
            "type": "string",
            "example": "correct-horse-battery"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "The refresh token issued by login, register, or a prior refresh."
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refresh_token": {
            "type": "string",
            "description": "The refresh token for the session being ended."
          }
        },
        "required": [
          "refresh_token"
        ]
      },
      "UpdateMeDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Rafi Hasan"
          },
          "class_level": {
            "type": "number",
            "example": 10,
            "description": "MVP supports 9 and 10 only (D-02)."
          },
          "curriculum_code": {
            "type": "string",
            "example": "nctb"
          },
          "medium": {
            "type": "string",
            "enum": [
              "BANGLA",
              "ENGLISH",
              "ENGLISH_VERSION"
            ]
          },
          "target_exam": {
            "type": "object",
            "example": "SSC",
            "nullable": true
          },
          "daily_goal_minutes": {
            "type": "number",
            "example": 45,
            "minimum": 15,
            "maximum": 240
          },
          "preferred_language": {
            "type": "string",
            "enum": [
              "bn",
              "en",
              "banglish"
            ]
          },
          "subject_ids": {
            "description": "Replaces the entire selection. Send the full list, not a delta — a student dropping a subject has to be expressible.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}